+++ src/common/collections/entities/collection.js
@@ -2,32 +2,24 @@
  * pwix:tenants-manager/src/common/collections/entities/collection.js
  */
 
+import { check, Match } from 'meteor/check';
 import { Logger } from 'meteor/pwix:logger';
 import { Mongo } from 'meteor/mongo';
 import { ReactiveVar } from 'meteor/reactive-var';
 import { Tracker } from 'meteor/tracker';
 
-import '../../js/index.js';
-
 const logger = Logger.get();
 
 export const Entities = {
-    collectionReady: new ReactiveVar( false ),
+    collectionName: null,
     fieldSet: new ReactiveVar( null )
 };
 
 Tracker.autorun(() => {
     const collectionName = TenantsManager.configure().tenantsCollection+'_e';
-    const ready = Entities.collectionReady.get();
-    if( collectionName && !ready ){
-        Entities.collectionName = collectionName;
-        Entities.collection = new Mongo.Collection( collectionName );
-        Entities.collectionReady.set( true );
-    }
-});
-
-Tracker.autorun(() => {
-    //logger.debug( 'Entities.collectionReady', Entities.collectionReady.get());
+    check( collectionName, Match.NonEmptyString );
+    Entities.collectionName = collectionName;
+    Entities.collection = new Mongo.Collection( collectionName );
 });
 
 TenantsManager.Entities = Entities;