+++ src/common/collections/records/collection.js
@@ -6,21 +6,16 @@ import { Mongo } from 'meteor/mongo';
 import { ReactiveVar } from 'meteor/reactive-var';
 import { Tracker } from 'meteor/tracker';
 
-import '../../js/index.js';
-
 export const Records = {
-    collectionReady: new ReactiveVar( false ),
+    collectionName: null,
     fieldSet: new ReactiveVar( null )
 };
 
 Tracker.autorun(() => {
     const collectionName = TenantsManager.configure().tenantsCollection+'_r';
-    const ready = Records.collectionReady.get();
-    if( collectionName && !ready ){
-        Records.collectionName = collectionName;
-        Records.collection = new Mongo.Collection( collectionName );
-        Records.collectionReady.set( true );
-    }
+    check( collectionName, Match.NonEmptyString );
+    Records.collectionName = collectionName;
+    Records.collection = new Mongo.Collection( collectionName );
 });
 
 TenantsManager.Records = Records;